+++ src/nlp/negation/README.md
@@ -1,0 +1,5 @@
+Modified ``NegSpacy`` to the needs of ``concept-graphs``  
+
+``NegSpacy`` in terms implemented the ``NegEx`` algorithm:  
+NegEx - A Simple Algorithm for Identifying Negated Findings and Diseases in Discharge Summaries.
+Chapman, Bridewell, Hanbury, Cooper, Buchanan https://doi.org/10.1006/jbin.2001.1029
