+++ src/nlp/negation/__init__.py
@@ -1,0 +1,6 @@
+"""Negation detection helpers for spaCy pipelines."""
+
+from src.nlp.negation.negation import Negex
+from src.nlp.negation.utils import FeaturesOfInterest
+
+__all__ = ["FeaturesOfInterest", "Negex"]
